chore #117 뷰모델 통합
