+++ Projects/Features/MainFeature/Sources/BroadcastCollectionViewModel.swift
@@ -32,10 +32,13 @@ public class BroadcastCollectionViewModel: ViewModel {
     public struct Input {
         let fetch: PassthroughSubject<Void, Never> = .init()
         let didTapStreamingDone: PassthroughSubject<Bool, Never> = .init()
+        let didWriteStreamingName: PassthroughSubject<String, Never> = .init()
     }
     
     public struct Output {
         let items: CurrentValueSubject<[Item], Never> = .init([])
+        let isActive: CurrentValueSubject<Bool, Never> = CurrentValueSubject(false)
+        let errorMessage: PassthroughSubject<String?, Never> = .init()
     }
     
     private let output = Output()
@@ -59,6 +62,15 @@ public class BroadcastCollectionViewModel: ViewModel {
             }
             .store(in: &cancellables)
         
+        input.didWriteStreamingName
+            .sink { [weak self] name in
+                guard let self else { return }
+                let validness = valid(name)
+                self.output.isActive.send(validness.isValid)
+                self.output.errorMessage.send(validness.errorMessage)
+            }
+            .store(in: &cancellables)
+        
         return output
     }
     
@@ -66,6 +78,24 @@ public class BroadcastCollectionViewModel: ViewModel {
         Task {
             let fetchedItems = await fetcher.fetch()
             output.items.send(fetchedItems)
+        }
+    }
+    
+    /// 방송 이름이 유효한지 확인하는 메서드
+    /// - Parameter _:  방송 이름
+    /// - Returns: (Bool, String?) - 유효 여부와 에러 메시지
+    private func valid(_ value: String) -> (isValid: Bool, errorMessage: String?) {
+        let isLengthValid = 3...20 ~= value.count
+        let isCharactersValid = value.allSatisfy { $0.isLetter || $0.isNumber || $0 == "_" }
+        
+        if !isLengthValid && !isCharactersValid {
+            return (false, "3글자 이상,20글자 이하로 입력해 주세요. 특수문자는 언더바(_)만 가능합니다.")
+        } else if !isLengthValid {
+            return (false, "최소 3글자 이상, 최대 20글자 이하로 입력해 주세요.")
+        } else if !isCharactersValid {
+            return (false, "특수문자는 언더바(_)만 가능합니다.")
+        } else {
+            return (true, nil)
         }
     }
 }
